+++ src/frontend/apps/impress/src/components/index.ts
@@ -3,7 +3,7 @@ export * from './Box';
 export * from './BoxButton';
 export * from './Card';
 export * from './DropButton';
-export * from './DropdownMenu';
+export * from './dropdown-menu/DropdownMenu';
 export * from './quick-search';
 export * from './Icon';
 export * from './InfiniteScroll';